app/requirements.txt
Flask
mysql-connector
python-dotenv